+++ src/resources/environments/environment.ts
@@ -81,10 +81,8 @@ export interface EnvironmentIncludes {
 }
 
 export interface EnvironmentMeta {
-  counts?: {
-    containers: StatefulCounts<ContainerState>;
-    instances: StatefulCounts<Instances.InstanceState>;
-  };
+  container_count?: StatefulCounts<ContainerState>;
+  instances_count?: StatefulCounts<Instances.InstanceState>;
   containers?: {
     id: ResourceId;
     name: string;